+++ src/core/homebridge-ipc/homebridge-ipc.service.ts
@@ -15,6 +15,7 @@ export class HomebridgeIpcService extends EventEmitter {
     'childBridgeMetadataResponse',
     'childBridgeStatusUpdate',
     'serverStatusUpdate',
+    'matterEvent',
   ]
 
   constructor(