+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - hide ssl settings when on hb raspbian image
 - added config option to configure a webroot for the ui
 - add a complete https management to settings (#2617) (@donavanbecker)
+- integration with updated homebridge version
 
 ### Other Changes
 
@@ -23,7 +24,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 
 - `@homebridge/hap-client` @ `v3.1.2`
 - `@homebridge/node-pty-prebuilt-multiarch` @ `v0.13.1`
-- `@homebridge/plugin-ui-utils` @ `v2.1.0`
+- `@homebridge/plugin-ui-utils` @ `v2.1.1-beta.0`
 
 ## v5.9.0 (2025-11-01)